Copy data from Bai3.4
