+++ Learning System/Learning System/AddNewQuestionForm.resx
@@ -61,7 +61,7 @@
   <data name="AddNewQuestionForm_ExpandPic.Image" type="System.Drawing.Bitmap, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
     <value>
         iVBORw0KGgoAAAANSUhEUgAAAgAAAAIACAYAAAD0eNT6AAAABGdBTUEAALGPC/xhBQAAAAlwSFlzAAAO
-        wQAADsEBuJFr7QAADfxJREFUeF7t1ltyHDcWRVFFe/5j8Oh6HN2W5BQp8RRZj3wAB2tF7B9bEouVwL35
+        vAAADrwBlbxySQAADfxJREFUeF7t1ltyHDcWRVFFe/5j8Oh6HN2W5BQp8RRZj3wAB2tF7B9bEouVwL35
         DQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
         AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
         AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
