--- Learning System/Learning System/App.config
@@ -1,15 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<configuration>
-    <configSections>
-        <sectionGroup name="applicationSettings" type="System.Configuration.ApplicationSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" >
-            <section name="Learning_System.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" requirePermission="false" />
-        </sectionGroup>
-    </configSections>
-    <applicationSettings>
-        <Learning_System.Properties.Settings>
-            <setting name="ChoosingContest" serializeAs="String">
-                <value>DefaultContest</value>
-            </setting>
-        </Learning_System.Properties.Settings>
-    </applicationSettings>
-</configuration>
