--- Learning System/Learning System/Category.json
@@ -1,158 +1,0 @@
-[
-  {
-    "Id": 0,
-    "Name": "Course: IT",
-    "SubArray": [
-      1,
-      5
-    ],
-    "QuestionArray": [],
-    "Description": "",
-    "IdNumber": "abf"
-    
-  },
-  {
-    "Id": 1,
-    "Name": "Lý",
-    "SubArray": [
-      2,
-      3,
-      4
-    ],
-    "QuestionArray": [
-      0,
-      1,
-      2,
-      3,
-      4
-    ],
-    "Description": "",
-    "IdNumber": "abf"
-    
-  },
-  {
-    "Id": 2,
-    "Name": "Lý Dễ",
-    "SubArray": [],
-    "QuestionArray": [
-      5,
-      6,
-      7,
-      8
-    ],
-    "Description": "",
-    "IdNumber": "abc"
-    
-  },
-  {
-    "Id": 3,
-    "Name": "Lý Vừa",
-    "SubArray": [],
-    "QuestionArray": [
-      11,
-      12,
-      13,
-      14,
-      15
-    ],
-    "Description": "",
-    "IdNumber": "abd"
-    
-  },
-  {
-    "Id": 4,
-    "Name": "Lý Khó",
-    "SubArray": [],
-    "QuestionArray": [
-      16,
-      17,
-      18,
-      19,
-      20,
-      21,
-      22,
-      23,
-      24,
-      25,
-      26,
-      27,
-      28
-    ],
-    "Description": "",
-    "IdNumber": "abe"
-    
-  },
-  {
-    "Id": 5,
-    "Name": "Sử",
-    "SubArray": [
-      6,
-      7,
-      8
-    ],
-    "QuestionArray": [
-      29,
-      30,
-      31,
-      32,
-      33
-    ],
-    "Description": "",
-    "IdNumber": "bcd"
-    
-  },
-  {
-    "Id": 6,
-    "Name": "Sử Dễ",
-    "SubArray": [],
-    "QuestionArray": [
-      34,
-      35,
-      36,
-      37,
-      38,
-      39
-    ],
-    "Description": "",
-    "IdNumber": "bce"
-    
-  },
-  {
-    "Id": 7,
-    "Name": "Sử Vừa",
-    "SubArray": [],
-    "QuestionArray": [
-      40,
-      41,
-      42,
-      43,
-      44
-    ],
-    "Description": "",
-    "IdNumber": "cde"
-    
-  },
-  {
-    "Id": 8,
-    "Name": "Sử Khó",
-    "SubArray": [],
-    "QuestionArray": [
-      45,
-      46,
-      47,
-      48,
-      49,
-      50,
-      51,
-      52,
-      53,
-      54,
-      55,
-      56,
-      57,
-      58
-    ],
-    "Description": "",
-    "IdNumber": "cdf"
-  }
-]
