+++ Learning System/Learning System/Class/Questions.cs
@@ -1,5 +1,4 @@
-using Newtonsoft.Json;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -14,8 +13,6 @@ namespace Learning_System
         public int CategoryID { get; set; }
         public string? Content { get; set; }
         public double DefaultMark { get; set; }
-
-        [JsonProperty("Choice")]
         public List<QuestionChoice> Choice { get; set; } = new List<QuestionChoice>();
     }
 
@@ -23,6 +20,5 @@ namespace Learning_System
     {
         public string? choice { get; set; }
         public double mark { get; set; }
-
     }
 }
