--- Learning System/Learning System/Contest.json
@@ -1,22 +1,0 @@
-[
-  {
-    "Name": "Bài kiểm tra Sử 12",
-    "Description": "",
-    "DescriptionShow": true,
-    "QuestionArray": [ 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58 ],
-    "ShuffleAnswer": false,
-    "TimeStart": "2023-06-05 11:00:00",
-    "TimeEnd": "2023-06-05 12:00:00",
-    "TimeLimit": 45
-  },
-  {
-    "Name": "Bài kiểm tra Lý 11",
-    "Description": "",
-    "DescriptionShow": true,
-    "QuestionArray": [ 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28 ],
-    "ShuffleAnswer": true,
-    "TimeStart": "2023-05-05 11:00:00",
-    "TimeEnd": "2023-05-05 12:00:00",
-    "TimeLimit": 45
-  }
-]
