+++ Learning System/Learning System/ExternalClass/DataProcessing.cs
@@ -1,6 +1,5 @@
-﻿// DataProcessing.cs v3.0.3
-// Last modified: 13.5.2023 by DH
-// Add ImportedColumns & Modified 2nd Import to load only selected columns
+// DataProcessing.cs v3.0.0
+// Last modified: 26.4.2023 by DH
 
 using Newtonsoft.Json.Linq;
 using System.Data;
@@ -10,7 +9,7 @@ namespace Learning_System.ExternalClass
     public class DataProcessing
     {
         private JArray ListElements { get; set; } = new JArray();
-
+        
         /// <summary>
         /// Lưu ListElements ngay trước đó. Cập nhật trước khi thực hiện hàm Get(), GetFirstRow(), Insert(), Update(), DeleteAll(), Delete()
         /// </summary>
@@ -28,8 +27,6 @@ namespace Learning_System.ExternalClass
             public const int Unset = 0;
         }
 
-        public bool ImportedColumns = false;
-
         /// <summary>
         /// Nhập dữ liệu (cột)
         /// </summary>
@@ -39,14 +36,12 @@ namespace Learning_System.ExternalClass
         /// <returns>StatusCode: OK (1): Thành công; Error (2): Thất bại</returns>
         public int Import(List<string> _columns, List<Type> _columnsType, List<string> _columnsKey)
         {
-            ImportedColumns = true;
-
             if (_columnsKey.Contains("PRIMARY KEY") == false)
             {
                 MessageBox.Show("Cần có ít nhất một cột có tính chất Primary (Unique + Not null) trong bảng!\n", "Error");
                 return StatusCode.Error;
             }
-
+            
             if (_columns.Count != _columnsType.Count || _columns.Count != _columnsKey.Count)
             {
                 MessageBox.Show("Số lượng các trường danh sách của cột không tương thích với nhau!\n", "Error");
@@ -67,7 +62,7 @@ namespace Learning_System.ExternalClass
         /// </summary>
         /// <param name="_jsonDataList">Danh sách dữ liệu đã xuất ra từ file JSON</param>
         /// <returns>StatusCode: OK (1): Thành công; Error (2): Thất bại</returns>
-        public int Import(JArray? _jsonDataList)
+        public int Import(JArray _jsonDataList)
         {
             try
             {
@@ -76,31 +71,7 @@ namespace Learning_System.ExternalClass
                     ListElements.Clear();
 
                     foreach (JObject _jsonObj in _jsonDataList.Cast<JObject>())
-                    {
-                        JObject _newObj = new();
-
-                        foreach (var _col in ColumnsSetting)
-                        {
-                            if ((_col.Key.Contains("NOT NULL") == true || _col.Key.Contains("PRIMARY KEY") == true) && _jsonObj[_col.Name] == null)
-                            {
-                                MessageBox.Show("Không thể nhập dữ liệu: " + _col.Name + " chứa giá trị NULL trong khi cột được đặt là NOT NULL!", "Error");
-                                return StatusCode.Error;
-                            }
-                            else if (_col.Key.Contains("PRIMARY KEY") == true || _col.Key.Contains("UNIQUE") == true)
-                            {
-                                foreach (var _row in ListElements)
-                                    if (_jsonObj[_col.Name].ToString().ToLower() == _row[_col.Name].ToString().ToLower())
-                                    {
-                                        MessageBox.Show("Không thể thêm phần tử mới: " + _col.Name + " chứa giá trị trùng lặp trong khi cột được đặt là UNIQUE!", "Error");
-                                        return StatusCode.Error;
-                                    }
-                            }
-
-                            _newObj.Add(_col.Name, _jsonObj[_col.Name]);
-                        }
-
-                        ListElements.Add(_newObj);
-                    }
+                        ListElements.Add(_jsonObj);
 
                     return StatusCode.OK;
                 }
@@ -152,10 +123,8 @@ namespace Learning_System.ExternalClass
                         break;
                     }
 
-                /*
                 if (_existPrimaryKeyOrUnique == false)
                     MessageBox.Show("Nên có ít nhất một cột PRIMARY KEY hoặc UNIQUE trong danh sách cột!", "Warning!", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning);
-                */
 
                 List<int> _choseIndex = new();
                 int length = ListElements.Count;
@@ -226,7 +195,7 @@ namespace Learning_System.ExternalClass
             {
                 List<int>? getIndex = GetAllSatisfy();
 
-                if (getIndex == null)
+                if (getIndex == null)    
                     return null;
 
                 DataTable _dataTable = new();
@@ -263,7 +232,7 @@ namespace Learning_System.ExternalClass
                 }
 
                 /// Add DataTable rows
-                foreach (int _index in getIndex)
+                foreach(int _index in getIndex)
                 {
                     DataRow _dataRow = _dataTable.NewRow();
 
@@ -372,7 +341,7 @@ namespace Learning_System.ExternalClass
         public DataRow? GetFirstRow()
         {
             DataTable? _dt = Get();
-
+            
             if (_dt == null)
                 return null;
             else
@@ -396,7 +365,7 @@ namespace Learning_System.ExternalClass
         {
             PrevListElements = ListElements;
 
-            if (_element == null)
+            if(_element == null)
             {
                 MessageBox.Show("Không thể thêm phần tử mới: Phần tử mới rỗng (null)!", "Error");
                 return StatusCode.Error;
@@ -423,6 +392,7 @@ namespace Learning_System.ExternalClass
                 }
 
                 ListElements.Add(_element);
+                MessageBox.Show("Đã thêm phần tử mới!", "Success", MessageBoxButtons.OK, MessageBoxIcon.Information);
                 return StatusCode.OK;
             }
             catch (Exception ex)
@@ -440,7 +410,7 @@ namespace Learning_System.ExternalClass
             PrevListElements = ListElements;
             ListElements.Clear();
         }
-        
+
         /// <summary>
         /// Xóa (các) bản ghi thỏa mãn điều kiện (kết hợp với Offset(), Limit(), Query(), Select(), Sort())
         /// </summary>
@@ -449,7 +419,7 @@ namespace Learning_System.ExternalClass
         {
             PrevListElements = ListElements;
 
-            if (this == null)
+            if(this == null)
                 return StatusCode.Error;
 
             try
@@ -463,6 +433,7 @@ namespace Learning_System.ExternalClass
                 foreach (var item in removeList)
                     ListElements.RemoveAt(item);
 
+                MessageBox.Show("Đã xóa thành công (các) phần tử!", "Success");
                 return StatusCode.OK;
             }
             catch (Exception ex)
@@ -481,10 +452,10 @@ namespace Learning_System.ExternalClass
         {
             PrevListElements = ListElements;
 
-            if (this == null)
+            if(this == null)
                 return StatusCode.Error;
 
-            if (_newValue == null)
+            if(_newValue == null)
             {
                 MessageBox.Show("Không thể cập nhật phần tử: Phần tử cập nhật rỗng!", "Error");
                 return StatusCode.Error;
@@ -493,7 +464,7 @@ namespace Learning_System.ExternalClass
             {
                 List<int>? updateList = GetAllSatisfy();
 
-                if (updateList == null)
+                if(updateList == null)
                     return StatusCode.Error;
 
                 foreach (var _col in ColumnsSetting)
@@ -510,7 +481,7 @@ namespace Learning_System.ExternalClass
                             {
                                 MessageBox.Show("Không thể cập nhật phần tử: " + _col.Name + " chứa giá trị trùng lặp trong khi cột được đặt là UNIQUE!", "Error");
                                 return StatusCode.Error;
-                            }
+                            }                                
                     }
                 }
 
@@ -518,6 +489,7 @@ namespace Learning_System.ExternalClass
                     foreach (JProperty jProperty in _newValue.Properties())
                         ListElements[item][jProperty.Name] = jProperty.Value;
 
+                MessageBox.Show("Đã cập nhật thành công (các) phần tử!", "Success");
                 return StatusCode.OK;
             }
             catch (Exception ex)
@@ -554,7 +526,7 @@ namespace Learning_System.ExternalClass
         /// </summary>
         public void Undo()
         {
-            ListElements = PrevListElements;
+             ListElements = PrevListElements;
         }
     }
 }
