--- Learning System/Learning System/Properties/Settings.settings
@@ -1,9 +1,0 @@
-﻿<?xml version='1.0' encoding='utf-8'?>
-<SettingsFile xmlns="http://schemas.microsoft.com/VisualStudio/2004/01/settings" CurrentProfile="(Default)" GeneratedClassNamespace="Learning_System.Properties" GeneratedClassName="Settings">
-  <Profiles />
-  <Settings>
-    <Setting Name="ChoosingContest" Type="System.String" Scope="Application">
-      <Value Profile="(Default)">DefaultContest</Value>
-    </Setting>
-  </Settings>
-</SettingsFile>
